requirements.txt
fastapi
uvicorn
pandas
python-dotenv
typer
kiteconnect
requests
google-generativeai